[android] Implemented subscription fragment delegate interface to separate and geralized ui logic between three different subscription screens
