+++ android/src/com/mapswithme/maps/purchase/BookmarksSightsSubscriptionFragment.java
@@ -9,12 +9,9 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.mapswithme.maps.R;
 
+@SuppressWarnings("WeakerAccess")
 public class BookmarksSightsSubscriptionFragment extends AbstractBookmarkSubscriptionFragment
 {
-  @SuppressWarnings("NullableProblems")
-  @NonNull
-  private SubscriptionFragmentDelegate mDelegate;
-
   @NonNull
   @Override
   PurchaseController<PurchaseCallback> createPurchaseController()
@@ -22,67 +19,26 @@ public class BookmarksSightsSubscriptionFragment extends AbstractBookmarkSubscri
     return PurchaseFactory.createBookmarksSightsSubscriptionController(requireContext());
   }
 
+  @NonNull
+  @Override
+  SubscriptionFragmentDelegate createFragmentDelegate(@NonNull AbstractBookmarkSubscriptionFragment fragment)
+  {
+    return new TwoButtonsSubscriptionFragmentDelegate(fragment);
+  }
+
   @Nullable
   @Override
   View onSubscriptionCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                                 @Nullable Bundle savedInstanceState)
   {
-    View root = inflater.inflate(R.layout.fragment_sightseeing_subscription, container, false);
-
-    mDelegate = new SubscriptionFragmentDelegate(this);
-    mDelegate.onSubscriptionCreateView(root);
-
-    return root;
+    return inflater.inflate(R.layout.fragment_sightseeing_subscription, container, false);
   }
 
-  @Override
-  void onSubscriptionDestroyView()
-  {
-    mDelegate.onSubscriptionDestroyView();
-  }
 
   @NonNull
   @Override
   SubscriptionType getSubscriptionType()
   {
     return SubscriptionType.BOOKMARKS_SIGHTS;
-  }
-
-  @Override
-  public void onProductDetailsLoading()
-  {
-    super.onProductDetailsLoading();
-    mDelegate.onProductDetailsLoading();
-  }
-
-  @Override
-  void hideButtonProgress()
-  {
-    mDelegate.hideButtonProgress();
-  }
-
-  @Override
-  void showButtonProgress()
-  {
-    mDelegate.showButtonProgress();
-  }
-
-  @NonNull
-  @Override
-  PurchaseUtils.Period getSelectedPeriod()
-  {
-    return mDelegate.getSelectedPeriod();
-  }
-
-  @Override
-  public void onReset()
-  {
-    mDelegate.onReset();
-  }
-
-  @Override
-  public void onPriceSelection()
-  {
-    mDelegate.onPriceSelection();
   }
 }
